+++ src/zeus_agent/goal_intelligence_runtime/__init__.py
@@ -1,0 +1,7 @@
+from zeus_agent.goal_intelligence_runtime.models import GoalIntelligenceContract
+from zeus_agent.goal_intelligence_runtime.runtime import build_goal_intelligence_contract
+
+__all__ = [
+    "GoalIntelligenceContract",
+    "build_goal_intelligence_contract",
+]